pyproject.toml
[project]
name = "atomic-habits-tracker"
version = "0.1.0"
package-mode = false
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "redis[async] (>=7.1.0,<8.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "drf-spectacular-sidecar (>=2025.12.1,<2026.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.18.2"
black = "^25.11.0"
isort = "^7.0.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

